+++ marshmallow/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
 
-__version__ = '1.0.1'
+__version__ = '1.1.0'
 __author__ = 'Steven Loria'
 __license__ = 'MIT'